refactor: move geometry utils to subfolder
